requirements.txt
python-dotenv
pypdf
google-genai
numpy